Gather agent event rendering into one module
